Added handlers for when the operator buttons are pressed.
Now the correct calculations are done when different operator buttons are pressed. The calculation can carry on from the answer in this calculator(ie: the answer you get is 12, then you click "-" and then you click 2, then it shows 10, so if you want to reset the calculation or do a new calculation you press AC to clear.